moving registrar and broker together
